// src/lib/schemas/rewardSchema.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Maree Backend API
 * OpenAPI spec version: 1.0.0
 */
import type { RewardSchemaDiscount } from "./rewardSchemaDiscount";

export interface RewardSchema {
  id: string;
  name: string;
  description: string;
  status: string;
  /** @pattern ^d+$ */
  cost: string;
  discountId: string;
  /** @nullable */
  image: string | null;
  createdAt: string;
  /** @nullable */
  deletedAt: string | null;
  discount: RewardSchemaDiscount;
}